Document `sequence` and `iterator` forgetting to run finalizers
Originally reported in KTIJ-36543 as an IDE inspection proposal.

When using `sequence {}` or `iterator {}` functions from
`kotlin.sequences`, there is a risk of not running `finally` blocks
and therefore, leaking resources if the consumer does not keep
asking for more values while the block is in a `try` section of
a `try`/`finally`.

As opposed to `kotlinx.coroutines` Flows, `sequence` and `iterator`
will not resume their code blocks with a cancellation exception
when they are no longer consumed, since they have no way of knowing
if they will be queried again at some later point.

It is obvious that the behavior couldn't have been different when
you consider the involved APIs carefully, but this is a pitfall
worth at least documenting.